refactor(ui): AppViewModel into commonMain, shared by App.kt and WebApp
AppViewModel (the cold-start gate: isInitialized / isLoggedIn / isBunkerVerifying
flows + an init{} that runs initialize() with the 30s forceInitialized fallback)
is pure, so move it to commonMain. Compose App.kt keeps using it via viewModel{}
(same package, unchanged); WebApp now consumes it via useViewModel and drops its
duplicated launchApp { withTimeoutOrNull(...) initialize } — the VM's init{} does
that on both platforms now.

Web-only shell concerns stay in WebApp's useEffectOnce (focus listeners, modal
focus trap, visibilitychange -> onBackground/onForeground).

The logged-in AppShell layout (relay/account/group/unread, ~40 repo refs) is a
separate, larger concern that spans many subsystems and has no single Compose
ViewModel counterpart (native composes App.kt + per-feature VMs); not in scope here.
